+++ launcher.py
@@ -1,5 +1,5 @@
 from lib.bot import bot
 
-VERSION = "2.0.7"
+VERSION = "2.0.8"
 
 bot.run(VERSION)